Add support for parallel native indexing with shuffle for perfect rollup.

+++ indexing-service/src/main/java/org/apache/druid/indexing/common/task/batch/parallel/TaskMonitor.java
@@ -181,30 +181,32 @@ public class TaskMonitor<T extends Task>
   public void stop()
   {
     synchronized (startStopLock) {
-      running = false;
-      taskStatusChecker.shutdownNow();
-
-      if (numRunningTasks > 0) {
-        final Iterator<MonitorEntry> iterator = runningTasks.values().iterator();
-        while (iterator.hasNext()) {
-          final MonitorEntry entry = iterator.next();
-          iterator.remove();
-          final String taskId = entry.runningTask.getId();
-          log.info("Request to kill subtask[%s]", taskId);
-          indexingServiceClient.killTask(taskId);
-          numRunningTasks--;
-          numKilledTasks++;
-        }
+      if (running) {
+        running = false;
+        taskStatusChecker.shutdownNow();
 
         if (numRunningTasks > 0) {
-          log.warn(
-              "Inconsistent state: numRunningTasks[%d] is still not zero after trying to kill all running tasks.",
-              numRunningTasks
-          );
-        }
-      }
+          final Iterator<MonitorEntry> iterator = runningTasks.values().iterator();
+          while (iterator.hasNext()) {
+            final MonitorEntry entry = iterator.next();
+            iterator.remove();
+            final String taskId = entry.runningTask.getId();
+            log.info("Request to kill subtask[%s]", taskId);
+            indexingServiceClient.killTask(taskId);
+            numRunningTasks--;
+            numKilledTasks++;
+          }
 
-      log.info("Stopped taskMonitor");
+          if (numRunningTasks > 0) {
+            log.warn(
+                "Inconsistent state: numRunningTasks[%d] is still not zero after trying to kill all running tasks.",
+                numRunningTasks
+            );
+          }
+        }
+
+        log.info("Stopped taskMonitor");
+      }
     }
   }
